@@ -1,8 +1,6 @@
-import 'package:meta/meta.dart';
-
 class Comment {
   final String name;
   final String content;
 
-  const Comment({@required this.name, @required this.content});
+  const Comment({required this.name, required this.content});
 }
