@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_reaction_button/flutter_reaction_button.dart';
-import 'data/example_data.dart' as Example;
 
+import 'data/example_data.dart' as Example;
 import 'items/item.dart';
 
 void main() => runApp(MyApp());
